+++ package.json
@@ -30,8 +30,6 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.7",
-    "node-sass": "^4.5.3",
-    "sass-loader": "^6.0.6",
     "style-loader": "^0.19.0",
     "volleyball": "^1.4.1",
     "webpack": "^3.8.1"
