+++ server/api/index.js
@@ -1,5 +1,5 @@
 const apiRouter = require('express').Router(); 
-const db = require('../../db/_db')
+const db = require('../db/_db')
 
 // If you aren't getting to this object, but rather the index.html (something with a joke) your path is wrong.
 	// I know this because we automatically send index.html for all requests that don't make sense in our backend.
