+++ server/db/_db.js
@@ -1,6 +1,6 @@
 const chalk = require('chalk');
 const Sequelize = require('sequelize');
-const pkg = require('../package.json');
+const pkg = require('../../package.json');
 
 console.log(chalk.yellow("Opening database connection"));
