+++ server/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 const path = require('path');
 const volleyball = require('volleyball');
 const bodyParser = require('body-parser');
-const db = require('../db/_db')
+const db = require('./db/_db')
 const PORT = 1337;
 const app = express();
